work in progress upload refactor

+++ app/src/main/java/de/westnordost/streetcomplete/data/osm/upload/OsmQuestsUpload.kt
@@ -1,5 +1,6 @@
 package de.westnordost.streetcomplete.data.osm.upload
 
+import android.os.CancellationSignal
 import android.util.Log
 import de.westnordost.osmapi.map.data.Element
 import de.westnordost.streetcomplete.ApplicationConstants.MAX_QUEST_UNDO_HISTORY_AGE
@@ -7,45 +8,56 @@ import javax.inject.Inject
 
 import de.westnordost.streetcomplete.data.QuestStatus
 import de.westnordost.streetcomplete.data.osm.OsmQuest
-import de.westnordost.streetcomplete.data.osm.persist.AOsmQuestDao
 import de.westnordost.streetcomplete.data.osm.persist.MergedElementDao
+import de.westnordost.streetcomplete.data.osm.persist.OsmQuestDao
+import de.westnordost.streetcomplete.data.upload.OnUploadedChangeListener
 import java.util.*
 
-open class OsmQuestsUpload @Inject constructor(
-    private val questDB: AOsmQuestDao,
+/** Gets all answered osm quests from local DB and uploads them via the OSM API */
+class OsmQuestsUpload @Inject constructor(
+    private val questDB: OsmQuestDao,
     private val elementDB: MergedElementDao,
-    private val singleOsmQuestChangeUpload: SingleOsmQuestUpload
+    private val changesetManager: OpenQuestChangesetsManager,
+    private val singleChangeUpload: SingleOsmElementTagChangesUpload
 ) {
     private val TAG = "OsmQuestUpload"
 
-	fun upload(changesetsUpload: OsmQuestChangesetsUpload) {
-		Log.i(TAG, "Applying quest changes")
+    var uploadedChangeListener: OnUploadedChangeListener? = null
 
-        val quests = questDB.getAll(null, QuestStatus.ANSWERED)
-		changesetsUpload.upload(quests, this::uploadSingle)
+    @Synchronized fun upload(signal: CancellationSignal) {
+        if (signal.isCanceled) return
+		Log.i(TAG, "Applying quest changes")
+        for (quest in questDB.getAll(null, QuestStatus.ANSWERED)) {
+            if (signal.isCanceled) break
+
+            try {
+                uploadSingle(quest)
+                quest.status = QuestStatus.CLOSED
+                questDB.update(quest)
+                Log.d(TAG, "Uploaded osm quest ${quest.toLogString()}")
+                uploadedChangeListener?.onUploaded()
+            } catch (e: ElementConflictException) {
+                Log.d(TAG, "Dropped osm quest ${quest.toLogString()}: ${e.message}")
+                /* #812 conflicting quests may not reside in the database, otherwise they would
+                   wrongfully be candidates for an undo - even though nothing was changed */
+                questDB.delete(quest.id!!)
+                uploadedChangeListener?.onDiscarded()
+            }
+        }
 
 		questDB.deleteAllClosed(System.currentTimeMillis() - MAX_QUEST_UNDO_HISTORY_AGE)
 	}
 
-    private fun uploadSingle(changesetId: Long, quest: OsmQuest) : List<Element> {
+    private fun uploadSingle(quest: OsmQuest) : Element {
         val element = elementDB.get(quest.elementType, quest.elementId)
+            ?: throw ElementDeletedException("Element deleted")
 
-        try {
-            if (element == null) throw ElementDeletedException("Element deleted")
-
-            val updatedElement = singleOsmQuestChangeUpload.upload(changesetId, quest, element)
-
-            quest.status = QuestStatus.CLOSED
-            questDB.update(quest)
-            return listOf(updatedElement)
-        }
-        catch (e: ElementConflictException) {
-            Log.d(TAG, "Dropping osm quest ${quest.toLogString()}: ${e.message}")
-            /* #812 conflicting quests may not reside in the database, otherwise they would
-               wrongfully be candidates for an undo - even though nothing was changed */
-            questDB.delete(quest.id!!)
-
-            throw e
+        return try {
+            val changesetId = changesetManager.getOrCreateChangeset(quest.osmElementQuestType, quest.changesSource!!)
+            singleChangeUpload.upload(changesetId, quest, element)
+        }  catch (e: ChangesetConflictException) {
+            val changesetId = changesetManager.createChangeset(quest.osmElementQuestType, quest.changesSource!!)
+            singleChangeUpload.upload(changesetId, quest, element)
         }
     }
 }
